pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.1",
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
